Drop subsystem.select param
The subsystem.select param in CS.cfg was used to record
whether the subsystem was installed as a new subsystem or
a clone of an existing subsystem, but it does not actually
control any functionality.

Ideally a clone should be indistinguishable from a normal
subsystem, so the param has been removed to minimize the
differences in the CS.cfg.

The pki-server status CLI has been modified to no longer
use the param. An upgrade script has also been added to
remove the param from existing instances.