Add test for CA with RSA algorithm
A new test has been added to install CA with a non-default RSA
algorithm and verify that the system and admin certs are created
with the correct key and signing algorithms. The test will also
issue an SSL server cert and verify that the cert will also be
created with the correct algorithm.

The tests for CA with ECC and RSA/PSS algorithms have also been
updated to perform the same validation.