Refactor RequestRepository.listRequestsByFilter()
Previously the RequestRepository.listRequestsByFilter() would
create a DBSession object, perform a search operation, close the
session, then return the DBSearchResults in a RequestList object
to the caller.

The problem is the DBSearchResults can only be used by the caller
as long as the database connection is still open, which is not
guaranteed once the session is closed.

To avoid potential issues, all variants of the method have been
updated to store the search results into a Collection while the
session is still active.