Lightweight CA read issuer certificate from nssdb
When working with HSM the Lightweight CA access the root CA certificate
from the CS.cfg. This has been modified reading the certificate from
NSSDB to allow the removal of such field from the configuration.

NOTE: Currently, Lightweight CA does not work with HSM so this change
cannot be really tested until the Github issue #2412 is fixed.