Add cert validity options/params for CLI and ACME
The pki nss-cert-issue command and the NSSIssuer in ACME have
been modified to provide options/params to specify the cert
validity in different units (e.g. minutes) which could be
useful for testing and end-users as well.

The old option/param is limited to months only so it has been
deprecated.