Fix Bug 2233158 - Make key wrapping algorithm configurable between AES-KWP and AES-CBC [RHCS 9.7.z].
This fix allows the TPS adminsitrator to configure the aesKeyWrap alg when the KRA wraps a private key to be injected onto the smart card.
This operation is done in server side keygen and key recovery scenarios. As of the latest code, only the AES_KWP wrap alg is supported.
Users requested the ability to choose either KWP or AES_CBC_PAD. The caveat here is that CBC_PAD is not assured compatibility between the token and certain hsm based setups. The default is to use the currently supported KWP alg. The reason to allow CBC is due to the fact that this alg is supported directly in the token and is much quicker than the hand coded KWP algorithm currently supported.

The choice is very simply configured in the TPS's CS.cfg and can be configured separately for various token profiles. Also the default is KWP if the following setting is not present:

ex: op.enroll.userKey.keyGen.aesKeyWrapAlg=CBC

The two choices here are "KWP" or "CBC".

Another example for recovery when the original token is marked as temporarilly lost:

op.enroll.userKeyTemporary.keyGen.aesKeyWrapAlg=CBC

Once this is set, subsequent calls from the TPS to the KRA for server side keygen will send this choice to the KRA.
From there the KRA will wrap the private key blob using the required alg and send it back to TPS.

From there the TPS will send the blob down to the coolkey applet to be unwrapped. A different byte code for KWP (0x88) or CBC (0x89) will be sent to a new coolkey applet. The new applet will be able to detect which alg to use and attempt.

There will be a subsequent addition to this PR to include the new applet which is still being finalized.

Once the user gets the new rpm's there will be a new applet to choose within the TPS config for ex:

op.format.userKey.update.applet.requiredVersion.prot.3=1.5.651f3902
op.enroll.userKey.update.applet.requiredVersion.prot.3=1.5.651f3902

The applet number here is just an example it will be different.

Review fixes...