Update KRA and OCSP tests
The test for basic KRA has been updated to check the security
domain and KRA connector in CA. The test for standalone KRA
has been updated to use a standalone CA so the CA should not
have a security domain and KRA connector.

Similarly, the test for basic OCSP has been updated to check
the security domain and OCSP publishing in CA. The test for
standalone OCSP has been updated to use a standalone CA as
well so the CA should not have a security domain and OCSP
publishing either.

Note: The KRA connector and the OCSP publishing can be added
later as a post-install task.