Restore support for cert bundle in pki_cert_chain_path
The NSSDatabase.import_cert_chain() has been modified
to support importing cert bundle by converting it into
a PKCS #7 file, then import it using the existing code.

The test for installing KRA on a separate instance has
been modified to use a root CA and a sub CA, then use
a cert bundle in the pki_cert_chain_path param.

The NSSCertImportCLI has been updated to avoid an NPE
if the nickname is not specified.

https://bugzilla.redhat.com/show_bug.cgi?id=2250162