Fix CRMFPopClient on FIPS environment
When FIPS is enabled RSA key generation cannot have both temporary and
sensitive to false.

Since temporary is defined from a command option the sensitive is set to
null so the default for the environment will be used (true if FIPS is
enabled and false if FIPS is disabled).

Fix BZ Bug 2250716