Add pki-server <subsystem>-db-repl-enable
The pki-server <subsystem>-db-repl-enable has been added
to enable replication on a single DS instance. This command
replaces the SubsystemDBReplicationSetupCLI which enables
replication both on the DS master and the replica at the
same time.

The test for CA cloning with replicated DS has been updated
to use the new command.