Fix the update to 11.5.0
Check and create the folder for every subsystem update, beside the
instance update.

During the update certs and csr are removed from CS.cfg file and stored
in `<config>/certs` folder if they are not in the internaldb.

The folder creation for the instance was done after update the subsystem generating an error and stopping the update.