Add tests for adding multiple LWCAs
New tests have been added to validate adding multiple LWCAs in
plain PKI environment and in IPA environment.

The CASigningUnit has been updated to no longer log an exception
if the cert or key is missing since in LWCA case the exception is
used as a mechanism to trigger the key retriever so it is actually
not a problem. Instead, the exception will only be logged by other
code where the exception indicates a real problem.