Update tests to use standard conf dir
Normally the Tomcat conf dir is located immediately under the
instance dir. For PKI server, which is based on Tomcat, the
location is /var/lib/pki/<instance>/conf.

Depending on the deployment scenario the config files might be
mounted from a different location (e.g. /etc/pki/<instance> on
Fedora) and the standard conf dir will be just a link to the
actual location.

To ensure that the tests are valid for all deployment scenarios
they have been updated to use the standard conf dir instead of
the Fedora-specific one.