Add test for CA with caDirPinUserCert profile
A new test has been added to set up the auth database using
the setpin tool, then perform PIN-authenticated enrollments
using the caDirPinUserCert profile.

The setpin tool has been modified to remove extra spaces in
the ACI attributes to make it easier to view and verify.

The pki ca-cert-request-submit command has been updated to
provide options to specify the enrollment password and PIN.

The CertRequestDAO.submitRequest() has been modified to get
the PIN from the enrollment request and store it into the
credentials object so that it can be authenticated later by
UidPwdPinDirAuthentication.