Update CA container
The CA container has been updated such that all files will be
owned by pkiuser:root so that it will work in Docker/Podman as
well as in OpenShift.

The test for CA container has been updated to use pki-server
commands to set up the CA database in the default DS backend
(i.e. userroot).