Add data folder for ACME container
The ACME container has been updated to support a data folder
which can be used to store the config and log files generated
by the server.

The ACME container has also been modified to store the default
config files created during the build. If the container is
started with an empty data folder the default config files will
be installed automatically.