Update CA container
The CA container has been modified to extend pki-server
instead of pki-runner such that it will be more consistent
with the ACME container.

The CA container has also been modified to store the default
config files created during the build. If the container is
started with an empty data folder the default config files
will be installed automatically.