Add serial and issuer to SSL logs and audits
When creating an SSL connection, logs and audits  where reporting only the
certificate subject. Since the certificate could be issued from other CAs
it could be difficult to identify.
This commit adds the issuer and the serial number of the certificate
in both the audit and log messages for a better identification.