Update PKI server container
The PKI server container (which is the base for all PKI
subsystem containers) has been updated to install the ROOT
webapp (which provides Web UI files) and PKI webapp (which
provides common PKI services), and store the default config
files.

The startup script has been modified to install the default
config files if it doesn't exist already, use PKI NSS CLI
instead of certutil to prepare the certs, and run the server
as pkiuser in Docker/Podman so that the log files will be
owned by pkiuser as well.

The test has been updated to check the files created by the
container and to verify that the container can be restarted
successfully.