Add test for CA container with existing config
A new test has been added to install a regular CA, then reuse
the certs, database, and config files to run the same CA as a
container.

The container startup scripts have been updated to provide
params to specify the nicknames of the existing certs and to
use password.conf to access the server's NSS database.