Move current REST APIs to v1 package
The current rest APIs are moved to a subpackage named `v1`. The new
implementation has moved to the subpackage `v1`. The new implementation
has moved to the subpackage `v2`. This will simplify the selective
building of REST implementation.

Subsystems ACME and EST have not been modified because they use REST to
implement the protocol and this requires extra effort.