Make ACL filter configurable for method and path
If the servlet has different ACLs for different HTTP methods or paths
the new implementation allows to define a map to identify the correct
ACL.

The MAP will associate a key with the ACL and the key is made as

     <method>:<path>

The method is one of the HTTP defined method (e.g. GET, POST, ...) and
the special value of "*" indicate all methods.

The path is internal of the servlet path and in case of path parameter
these can be identified with "{}". An example of entry could be:

     key= PUT:/token/{} value= token.read