Add TPS container
A new container has been added to provide a basic TPS
subsystem without any connectors. The connectors need
to be set up after the container is created. This is
necessary to allow creating clones of the container
without creating duplicate connectors.

pkispawn has been updated such that it will only set up
the connectors and the shared secret if the URLs to the
CA, KRA, and TKS are provided.

A new test has been added to create the initial CA, KRA,
TKS, and TPS containers. In the future the test will be
updated to set up the connectors and the shared secret,
and then test the token format and enroll operations.