Fix CMake build for missing libraries
The test for jackson and resteasy library was failing because the
variable is undefined when the libraries are not provided with the
package.
Since CMake test works with variables considering false when the
variable is undefined the all the is statements have been modified to
check if the variable is defined.

https://cmake.org/cmake/help/latest/command/if.html#variable