Add routing based the annotation WebAction
Using raw http servlet with base method it is not always possible to handle the
correct return state because the value depends on the operation of other
operation.

Using the annotation it is possible to map all the operations during the
init phase and provide the correct state in any situation.

Additionally, it is possible to split multiple operations associated to
a single http method simplifying the code.