Fix routing for duplicated path
In case where the same combination of "method:path" match the routing
it was selected always the first.

E.g.: POST:{} and POST:retrieve from KRA `KeyServlet` was not working
because always the first was used.

Additionally, the routing of servlets and filters are the same and it is
not supported anymore filter routing to multiple method using "*" which
has never been used.