Drop ClonesConnectivyAndDataCheck from pki-healthcheck
pki-healthcheck is a tool to check the status of an instance or
a node in a cluster so that if it reports a problem the admin can
fix it or the monitoring service can replace it with a new node.

The ClonesConnectivyAndDataCheck on the other hand is a plugin
that checks the connectivity from the instance to other clones.
This plugin will report a problem if another clone is down even
though the instance itself is fine, which would be misleading.
Since it doesn't really fit the purpose of pki-healthcheck this
plugin has been dropped.