Add pki ca-cert-issue
The pki ca-cert-request-submit command can be used in two ways.
If it's invoked with an install token, the cert will be issued
immediately. If it's invoked without an install token, it will
submit the request to the CA, but then the request will need to
be approved, and the cert will need to be retrieved separately.

To make it easier to issue a cert, a new pki ca-cert-issue has
been added which is similar to pki ca-cert-request-submit but
it can approve the request and retrieve the cert immediately if
invoked with the proper credentials.

pkispawn and most CI tests have been updated to use the new
command. The pki ca-cert-request-submit options that take an
install token have been deprecated.