Update rpminspect test
The pom.xml files have been updated to define the target
Java version using maven.compiler.release property. The
maven-compiler-plugin is no longer used so it has been
removed.

The pki-rpminspect.yaml has been updated to no longer
define the javabytecode requirement so it will use the
standard requirement for the platform.

The rpminspect test has been updated to call rpminspect
directly in separate steps to make it easier to inspect
the failures. The rpminspect.sh is no longer used so it
has been removed.

Currently the test is failing because some dependencies
were built with older Java bytecode versions. They need
to be rebuilt with the proper version.