Update test for CA with sequential serial numbers
pkispawn has been modified to provide more params to configure the
sequential serial numbers in CA. The params can also be added for
KRA if needed later.

The test for CA with sequential serial numbers has been updated to
perform more detailed steps and verification. The test will now use
small ranges to make it easier to verify the changes in the CS.cfg
and DS. The test will also use SerialNumberUpdateJob to update the
ranges immediately instead of waiting for SerialNumberUpdateTask
to run.

A more complex test with CA clones will be added separately later.