Add option for decimal range numbers
With legacy serial ID, request and replica ranges are expressed in decimal while serial in hex.
However, there is a problem in the hex management which create gaps in
the sequences every time a new range is allocated.

Since gaps could create problems to third party software it has been
introduced a new parameter in CS.cfg to set the same format for all ranges:

dbs.numberRangeRadix

If this is not present or negative then the current default is adopted
and nothing change.

If this is set to 10 all the values are handled as decimal and the
ranges work properly. If set to different values than 10 (e.g. 16 for
hex) the gap problem is still present.

This new parameter is not update during the update to avoid creating
problem to running instances which do not have problem with range gaps.

In case to move the value to 10 and solve the gaps problem then the
CS.cfg serial range has to be fixed considering the current value as
hex. Also the nextRange in the following DS node has to fixed
accordingly:

ou=certificateRepository,ou=ca,<prefix>