Add test for CA clone with sequential serial numbers
A new test has been added to create a CA with sequential serial
numbers, create a clone with the same config, perform enrollments,
and update the serial number ranges with pki ca-job-start
serialNumberUpdate.

Ideally the serial numbers should be contiguous, but currently
the code creates a gap between the ranges. The steps for fixing
an existing gap and migrating to RSNv3 may be added later.